Move ProductSortSelect and ProductListItem under /product-list
