+++ assets/js/base/components/product-list/product-list-item/index.js
